+++ src/FederationLib/Methods/Entities/DeleteEntity.php
@@ -61,10 +61,9 @@
                 EntitiesManager::deleteEntity($entityRecord->getUuid());
 
                 AuditLogManager::createEntry(AuditLogType::ENTITY_DELETED, sprintf(
-                    'Entity %s deleted by %s (%s)',
-                    $entityRecord->getUuid(),
-                    $authenticatedOperator->getName(),
-                    $authenticatedOperator->getUuid()
+                    'Entity %s deleted by %s',
+                    $entityRecord->getAddress(),
+                    $authenticatedOperator->getName()
                 ), $authenticatedOperator->getUuid());
             }
             catch (DatabaseOperationException $e)